Name the root route to match the store_url helper
The app’s routes are configured to route the root of the application to
`StoreController`, but this route wasn’t given a proper name of `store`
to match the `store_url` helper used elsewhere (like in
`CartsController` and `LineItemsController`).

To fix this problem, simply add the `:as` option to the `root` route
with the `store` value. See the [*Rails Routing from the Outside
In* guide][rg] for more information.

[rg]: http://guides.rubyonrails.org/routing.html#naming-routes